update valid / fix fe

+++ JewelrySalesSystem_NoName_FE/Pages/Manager/Materials/DeleteMaterial.cshtml.cs
@@ -4,6 +4,7 @@ using JewelrySalesSystem_NoName_FE.Ultils;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace JewelrySalesSystem_NoName_FE.Pages.Manager.Materials
 {
@@ -21,8 +22,17 @@ namespace JewelrySalesSystem_NoName_FE.Pages.Manager.Materials
 
         public async Task<IActionResult> OnGetAsync(Guid id)
         {
+            var token = HttpContext.Session.GetString("Token");
+            if (string.IsNullOrEmpty(token))
+            {
+                TempData["ErrorMessage"] = "Bạn cần phải login trước.";
+                return RedirectToPage("/Auth/Login");
+            }
+
             var apiUrl = $"{ApiPath.MaterialList}/id?id={id}";
-            var client = _httpClientFactory.CreateClient();
+            var client = _httpClientFactory.CreateClient("ApiClient");
+            client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
+
             var response = await client.GetStringAsync(apiUrl);
             Material = JsonConvert.DeserializeObject<MaterialDTO>(response);
 
@@ -36,8 +46,16 @@ namespace JewelrySalesSystem_NoName_FE.Pages.Manager.Materials
 
         public async Task<IActionResult> OnPostAsync(Guid id)
         {
+            var token = HttpContext.Session.GetString("Token");
+            if (string.IsNullOrEmpty(token))
+            {
+                TempData["ErrorMessage"] = "Bạn cần phải login trước.";
+                return RedirectToPage("/Auth/Login");
+            }
+
             var apiUrl = $"{ApiPath.MaterialList}/id?id={id}";
-            var client = _httpClientFactory.CreateClient();
+            var client = _httpClientFactory.CreateClient("ApiClient");
+            client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", token);
 
             var response = await client.DeleteAsync(apiUrl);
